Start of refactoring of trafficLight
